Validation
Rules and feedback, returning the properly status code, using accept application/json, using validation within model class and how to manage PATCH method to interate with not all validation rules.

+++ app/Http/Controllers/MarcaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Marca;
 use Illuminate\Http\Request;
+use Barryvdh\Debugbar\Facades\Debugbar;
 
 class MarcaController extends Controller
 {
@@ -19,7 +20,7 @@ class MarcaController extends Controller
     public function index()
     {
        $marca = $this->marca->all();
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -30,8 +31,13 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
+       
+
+       
+
+        $request->validate($this->marca->rules(), $this->marca->feedback());
         $brand = $this->marca->create($request->all());
-        return $brand;
+        return response()->json($brand, 201);
     }
 
     /**
@@ -42,9 +48,13 @@ class MarcaController extends Controller
      */
     public function show($id)
     {
-       $marca = $this->marca->find($id);
+       $brand = $this->marca->find($id);
+       if($brand === null){
+           return response()->json(['error' => "This brand doesn't extist"], 404); 
+       }
+       
 
-        return $marca;
+        return response()->json($brand, 200);
     }
 
     /**
@@ -57,9 +67,24 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['error' => "This brand doesn't extist"], 404); 
+        }
+        if($request->method() === 'PATCH'){
+            // allows update register using patch method
+            $dynamicRules = array();
+
+            foreach ($marca->rules() as $input => $rule) {
+                if(array_key_exists($input, $request->all())){
+                    $dynamicRules[$input] = $rule;                    
+                }                                    
+            }
+            $request->validate($dynamicRules, $marca->feedback());
+        } else {
+            $request->validate($marca->rules(), $marca->feedback());
+        }        
         $marca->update($request->all());
-        
-        return response()->json($marca);
+        return response()->json($marca, 200);
     }
 
     /**
@@ -71,7 +96,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return response()->json(['error' => "This brand doesn't extist"], 404); 
+        }
         $marca->delete();
-        return ['msg' => 'Marca removida com sucesso.'];
+        return response()->json(['msg' => 'Marca removida com sucesso.'], 200);
     }
 }
